+++ src/classes/queue-events.ts
@@ -47,10 +47,18 @@ export interface QueueEventsListener extends IoredisListener {
 
   /**
    * Listen to 'debounced' event.
+   * @deprecated use deduplicated event
    *
    * This event is triggered when a job is debounced because debounceId still existed.
    */
   debounced: (args: { jobId: string; debounceId: string }, id: string) => void;
+
+  /**
+   * Listen to 'deduplicated' event.
+   *
+   * This event is triggered when a job is deduplicated because deduplicatedId still existed.
+   */
+  deduplicated: (args: { jobId: string; deduplicationId: string }, id: string) => void;
 
   /**
    * Listen to 'delayed' event.