+++ src/types/job-options.ts
@@ -3,8 +3,14 @@ import { BaseJobOptions, DebounceOptions } from '../interfaces';
 export type JobsOptions = BaseJobOptions & {
   /**
    * Debounce options.
+   * @deprecated use deduplication option
    */
   debounce?: DebounceOptions;
+
+  /**
+   * Deduplication options.
+   */
+  deduplication?: DebounceOptions;
 
   /**
    * Modes when a child fails: fail, ignore, remove, wait.