+++ tests/test_child-pool.ts
@@ -2,6 +2,7 @@ import { expect } from 'chai';
 import { ChildPool } from '../src/classes';
 import { join } from 'path';
 
+const NoopProc = () => {};
 describe('Child pool for Child Processes', () => {
   sandboxProcessTests();
 });
@@ -32,70 +33,70 @@ function sandboxProcessTests(
 
     it('should return same child if free', async () => {
       const processor = __dirname + '/fixtures/fixture_processor_bar.js';
-      const child = await pool.retain(processor);
+      const child = await pool.retain(processor, NoopProc);
       expect(child).to.be.ok;
       pool.release(child);
       expect(pool.retained).to.be.empty;
-      const newChild = await pool.retain(processor);
+      const newChild = await pool.retain(processor, NoopProc);
       expect(child).to.be.eql(newChild);
     });
 
     it('should return a new child if reused the last free one', async () => {
       const processor = __dirname + '/fixtures/fixture_processor_bar.js';
-      let child = await pool.retain(processor);
+      let child = await pool.retain(processor, NoopProc);
       expect(child).to.be.ok;
       pool.release(child);
       expect(pool.retained).to.be.empty;
-      let newChild = await pool.retain(processor);
+      let newChild = await pool.retain(processor, NoopProc);
       expect(child).to.be.eql(newChild);
       child = newChild;
-      newChild = await pool.retain(processor);
+      newChild = await pool.retain(processor, NoopProc);
       expect(child).not.to.be.eql(newChild);
     });
 
     it('should return a new child if none free', async () => {
       const processor = __dirname + '/fixtures/fixture_processor_bar.js';
-      const child = await pool.retain(processor);
+      const child = await pool.retain(processor, NoopProc);
       expect(child).to.be.ok;
       expect(pool.retained).not.to.be.empty;
-      const newChild = await pool.retain(processor);
+      const newChild = await pool.retain(processor, NoopProc);
       expect(child).to.not.be.eql(newChild);
     });
 
     it('should return a new child if killed', async () => {
       const processor = __dirname + '/fixtures/fixture_processor_bar.js';
-      const child = await pool.retain(processor);
+      const child = await pool.retain(processor, NoopProc);
       expect(child).to.be.ok;
       await pool.kill(child);
       expect(pool.retained).to.be.empty;
-      const newChild = await pool.retain(processor);
+      const newChild = await pool.retain(processor, NoopProc);
       expect(child).to.not.be.eql(newChild);
     });
 
     it('should return a new child if many retained and none free', async () => {
       const processor = __dirname + '/fixtures/fixture_processor_bar.js';
       const children = await Promise.all([
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
       ]);
       expect(children).to.have.length(6);
-      const child = await pool.retain(processor);
+      const child = await pool.retain(processor, NoopProc);
       expect(children).not.to.include(child);
     }).timeout(10000);
 
     it('should return an old child if many retained and one free', async () => {
       const processor = __dirname + '/fixtures/fixture_processor_bar.js';
       const children = await Promise.all([
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
-        pool.retain(processor),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
+        pool.retain(processor, NoopProc),
       ]);
 
       expect(children).to.have.length(6);
@@ -108,7 +109,7 @@ function sandboxProcessTests(
       const processor = __dirname + '/fixtures/fixture_processor_bar.js';
       process.execArgv.push('--no-warnings');
 
-      const child = await pool.retain(processor);
+      const child = await pool.retain(processor, NoopProc);
       expect(child).to.be.ok;
       if (!useWorkerThreads) {
         expect(child.childProcess.spawnargs).to.include('--no-warnings');